Move routing parameters into constants.
