+++ console/src/app/Routing/params.routes.ts
@@ -1,0 +1,3 @@
+export class RouteParams { 
+    public static readonly ID = "id"; 
+}
