+++ console/src/app/Routing/routes.ts
@@ -1,13 +1,14 @@
 import { Routes } from "@angular/router";
-import { MongooseSetUpComponent } from './mongoose-set-up/mongoose-set-up.component';
-import { ConfigurationEditingRootComponent } from './mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component';
-import { RunsTableTabsComponent } from './runs-table/runs-table-tabs/runs-table-tabs.component';
-import { NodesComponent } from './mongoose-set-up/set-up-steps/nodes/nodes.component';
-import { ScenariosComponent } from './mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component';
-import { RunStatisticsComponent } from './run-statistics/run-statistics.component';
+import { MongooseSetUpComponent } from '../mongoose-set-up/mongoose-set-up.component';
+import { ConfigurationEditingRootComponent } from '../mongoose-set-up/set-up-steps/configuration-set-up/control-editing-root/control-editing-root.component';
+import { RunsTableTabsComponent } from '../runs-table/runs-table-tabs/runs-table-tabs.component';
+import { NodesComponent } from '../mongoose-set-up/set-up-steps/nodes/nodes.component';
+import { ScenariosComponent } from '../mongoose-set-up/set-up-steps/scenarios-set-up/scenarios/scenarios.component';
+import { RunStatisticsComponent } from '../run-statistics/run-statistics.component';
 import { RoutesList } from "./routes-list";
-import { RunStatisticLogsComponent } from "./run-statistics/run-statistic-logs/run-statistic-logs.component";
-import { RunStatisticsChartsComponent } from "./run-statistics/run-statistics-charts/run-statistics-charts.component";
+import { RunStatisticLogsComponent } from "../run-statistics/run-statistic-logs/run-statistic-logs.component";
+import { RunStatisticsChartsComponent } from "../run-statistics/run-statistics-charts/run-statistics-charts.component";
+import { RouteParams } from "./params.routes";
 
 
 export const routes: Routes = [
@@ -15,7 +16,7 @@ export const routes: Routes = [
 
     { path: RoutesList.RUNS, component: RunsTableTabsComponent },
 
-    { path: RoutesList.RUN_STATISTICS + '/:id', component: RunStatisticsComponent,
+    { path: RoutesList.RUN_STATISTICS + '/:' + RouteParams.ID, component: RunStatisticsComponent,
       children: [
         { path: RoutesList.RUN_LOGS, component: RunStatisticLogsComponent},
         { path: RoutesList.RUN_CHARTS, component: RunStatisticsChartsComponent}
